feat: add more Msgs and reorganize files
