+++ packages/injective-cosmwasm/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "injective-cosmwasm"
-version = "0.1.10"
+version = "0.1.12"
 authors = ["Albert Chon <albert@injectivelabs.org>"]
 edition = "2018"
 description = "Bindings for CosmWasm contracts to call into custom modules of Injective Core"
