+++ packages/injective-cosmwasm/src/msg.rs
@@ -1,8 +1,7 @@
-use injective_math::FPDecimal;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::route::InjectiveRoute;
+use crate::{derivative::DerivativeOrder, order::OrderData, route::InjectiveRoute, spot::SpotOrder};
 use cosmwasm_std::{Addr, Coin, CosmosMsg, CustomMsg};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -20,96 +19,56 @@ impl From<InjectiveMsgWrapper> for CosmosMsg<InjectiveMsgWrapper> {
 
 impl CustomMsg for InjectiveMsgWrapper {}
 
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct OrderData {
-    pub market_id: String,
-    pub subaccount_id: String,
-    pub order_hash: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct OrderInfo {
-    pub subaccount_id: String,
-    pub fee_recipient: String,
-    pub price: FPDecimal,
-    pub quantity: FPDecimal,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct SpotOrder {
-    pub market_id: String,
-    pub order_info: OrderInfo,
-    pub order_type: i32,
-    pub trigger_price: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct DerivativeOrder {
-    pub market_id: String,
-    pub order_info: OrderInfo,
-    pub order_type: i32,
-    pub margin: FPDecimal,
-    pub trigger_price: Option<String>,
-}
-
-impl DerivativeOrder {
-    pub fn new(
-        price: FPDecimal,
-        quantity: FPDecimal,
-        margin: FPDecimal,
-        is_buy: bool,
-        market_id: &str,
-        subaccount_id: &str,
-        fee_recipient: &str,
-    ) -> DerivativeOrder {
-        DerivativeOrder {
-            market_id: market_id.to_string(),
-            order_info: OrderInfo {
-                subaccount_id: subaccount_id.to_string(),
-                fee_recipient: fee_recipient.to_string(),
-                price,
-                quantity,
-            },
-            order_type: if is_buy { 1 } else { 2 }, // TODO PO-orders
-            margin,
-            trigger_price: None,
-        }
-    }
-    pub fn is_reduce_only(&self) -> bool {
-        self.margin.is_zero()
-    }
-    pub fn get_price(&self) -> FPDecimal {
-        self.order_info.price
-    }
-    pub fn get_qty(&self) -> FPDecimal {
-        self.order_info.quantity
-    }
-    pub fn get_val(&self) -> FPDecimal {
-        self.get_price() * self.get_qty()
-    }
-    pub fn get_margin(&self) -> FPDecimal {
-        self.margin
-    }
-    pub fn non_reduce_only_is_invalid(&self) -> bool {
-        self.get_margin().is_zero() || self.get_price().is_zero() || self.get_qty().is_zero()
-    }
-    pub fn reduce_only_price_is_invalid(&self) -> bool {
-        self.get_price().is_zero()
-    }
-    pub fn reduce_only_qty_is_invalid(&self) -> bool {
-        self.get_qty().is_zero()
-    }
-}
-
 /// InjectiveMsg is an override of CosmosMsg::Custom to add support for Injective's custom message types
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum InjectiveMsg {
+    Deposit {
+        sender: String,
+        subaccount_id: String,
+        amount: Coin,
+    },
+    Withdraw {
+        sender: String,
+        subaccount_id: String,
+        amount: Coin,
+    },
     SubaccountTransfer {
         sender: Addr,
         source_subaccount_id: String,
         destination_subaccount_id: String,
         amount: Coin,
+    },
+    ExternalTransfer {
+        sender: String,
+        source_subaccount_id: String,
+        destination_subaccount_id: String,
+        amount: Coin,
+    },
+    CreateSpotMarketOrder {
+        sender: String,
+        order: SpotOrder,
+    },
+    CreateDerivativeMarketOrder {
+        sender: String,
+        order: DerivativeOrder,
+    },
+    IncreasePositionMargin {
+        sender: String,
+        source_subaccount_id: String,
+        destination_subaccount_id: String,
+        market_id: String,
+        amount: Coin,
+    },
+    LiquidatePosition {
+        sender: String,
+        subaccount_id: String,
+        market_id: String,
+        order: Option<DerivativeOrder>,
+    },
+    RegisterAsDMM {
+        sender: String,
+        dmm_account: String,
     },
     BatchUpdateOrders {
         sender: String,
@@ -121,15 +80,30 @@ pub enum InjectiveMsg {
         spot_orders_to_create: Vec<SpotOrder>,
         derivative_orders_to_create: Vec<DerivativeOrder>,
     },
-    CreateDerivativeMarketOrder {
-        sender: String,
-        order: DerivativeOrder,
-    },
-    Deposit {
-        sender: String,
-        subaccount_id: String,
-        amount: Coin,
-    },
+}
+
+pub fn create_deposit_msg(sender: String, subaccount_id: String, amount: Coin) -> CosmosMsg<InjectiveMsgWrapper> {
+    InjectiveMsgWrapper {
+        route: InjectiveRoute::Exchange,
+        msg_data: InjectiveMsg::Deposit {
+            sender,
+            subaccount_id,
+            amount,
+        },
+    }
+    .into()
+}
+
+pub fn create_withdraw_msg(sender: String, subaccount_id: String, amount: Coin) -> CosmosMsg<InjectiveMsgWrapper> {
+    InjectiveMsgWrapper {
+        route: InjectiveRoute::Exchange,
+        msg_data: InjectiveMsg::Withdraw {
+            sender,
+            subaccount_id,
+            amount,
+        },
+    }
+    .into()
 }
 
 pub fn create_subaccount_transfer_msg(
@@ -146,6 +120,86 @@ pub fn create_subaccount_transfer_msg(
             destination_subaccount_id,
             amount,
         },
+    }
+    .into()
+}
+
+pub fn create_external_transfer_msg(
+    sender: String,
+    source_subaccount_id: String,
+    destination_subaccount_id: String,
+    amount: Coin,
+) -> CosmosMsg<InjectiveMsgWrapper> {
+    InjectiveMsgWrapper {
+        route: InjectiveRoute::Exchange,
+        msg_data: InjectiveMsg::ExternalTransfer {
+            sender,
+            source_subaccount_id,
+            destination_subaccount_id,
+            amount,
+        },
+    }
+    .into()
+}
+
+pub fn create_spot_market_order_msg(sender: String, order: SpotOrder) -> CosmosMsg<InjectiveMsgWrapper> {
+    InjectiveMsgWrapper {
+        route: InjectiveRoute::Exchange,
+        msg_data: InjectiveMsg::CreateSpotMarketOrder { sender, order },
+    }
+    .into()
+}
+
+pub fn create_derivative_market_order_msg(sender: String, order: DerivativeOrder) -> CosmosMsg<InjectiveMsgWrapper> {
+    InjectiveMsgWrapper {
+        route: InjectiveRoute::Exchange,
+        msg_data: InjectiveMsg::CreateDerivativeMarketOrder { sender, order },
+    }
+    .into()
+}
+
+pub fn create_increase_position_margin_msg(
+    sender: String,
+    source_subaccount_id: String,
+    destination_subaccount_id: String,
+    market_id: String,
+    amount: Coin,
+) -> CosmosMsg<InjectiveMsgWrapper> {
+    InjectiveMsgWrapper {
+        route: InjectiveRoute::Exchange,
+        msg_data: InjectiveMsg::IncreasePositionMargin {
+            sender,
+            source_subaccount_id,
+            destination_subaccount_id,
+            market_id,
+            amount,
+        },
+    }
+    .into()
+}
+
+pub fn create_liquidate_position_msg(
+    sender: String,
+    subaccount_id: String,
+    market_id: String,
+    order: Option<DerivativeOrder>,
+) -> CosmosMsg<InjectiveMsgWrapper> {
+    InjectiveMsgWrapper {
+        route: InjectiveRoute::Exchange,
+        msg_data: InjectiveMsg::LiquidatePosition {
+            sender,
+            subaccount_id,
+            market_id,
+            order,
+        },
+    }
+    .into()
+}
+
+pub fn create_register_as_dmm_msg(sender: String, dmm_account: String) -> CosmosMsg<InjectiveMsgWrapper> {
+    InjectiveMsgWrapper {
+        route: InjectiveRoute::Exchange,
+        msg_data: InjectiveMsg::RegisterAsDMM { sender, dmm_account },
     }
     .into()
 }
@@ -171,26 +225,6 @@ pub fn create_batch_update_orders_msg(
             derivative_orders_to_cancel,
             spot_orders_to_create,
             derivative_orders_to_create,
-        },
-    }
-    .into()
-}
-
-pub fn create_derivative_market_order_msg(sender: String, order: DerivativeOrder) -> CosmosMsg<InjectiveMsgWrapper> {
-    InjectiveMsgWrapper {
-        route: InjectiveRoute::Exchange,
-        msg_data: InjectiveMsg::CreateDerivativeMarketOrder { sender, order },
-    }
-    .into()
-}
-
-pub fn create_deposit_msg(sender: String, subaccount_id: String, amount: Coin) -> CosmosMsg<InjectiveMsgWrapper> {
-    InjectiveMsgWrapper {
-        route: InjectiveRoute::Exchange,
-        msg_data: InjectiveMsg::Deposit {
-            sender,
-            subaccount_id,
-            amount,
         },
     }
     .into()
