+++ packages/injective-cosmwasm/src/order.rs
@@ -1,0 +1,18 @@
+use injective_math::FPDecimal;
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct OrderData {
+    pub market_id: String,
+    pub subaccount_id: String,
+    pub order_hash: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct OrderInfo {
+    pub subaccount_id: String,
+    pub fee_recipient: String,
+    pub price: FPDecimal,
+    pub quantity: FPDecimal,
+}
