+++ packages/injective-cosmwasm/src/query.rs
@@ -2,7 +2,10 @@ use injective_math::FPDecimal;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::{msg::OrderInfo, route::InjectiveRoute};
+use crate::{
+    derivative::EffectivePosition, derivative::TrimmedDerivativeLimitOrder, derivative_market::FullDerivativeMarket,
+    derivative_market::PerpetualMarketFunding, derivative_market::PerpetualMarketInfo, exchange::Deposit, route::InjectiveRoute,
+};
 use cosmwasm_std::CustomQuery;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -20,6 +23,13 @@ pub enum InjectiveQuery {
     SubaccountDeposit {
         subaccount_id: String,
         denom: String,
+    },
+    SpotMarket {
+        market_id: String,
+    },
+    TraderSpotOrders {
+        market_id: String,
+        subaccount_id: String,
     },
     // DerivativeMarket will return the derivative market for a given id
     DerivativeMarket {
@@ -61,82 +71,9 @@ pub struct SubaccountDepositResponse {
     pub deposits: Deposit,
 }
 
-#[allow(non_snake_case)]
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct Position {
-    #[serde(default)]
-    pub isLong: bool,
-    pub quantity: FPDecimal,
-    pub entry_price: FPDecimal,
-    #[serde(default)]
-    pub margin: FPDecimal,
-    pub cumulative_funding_entry: FPDecimal,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct EffectivePosition {
-    #[serde(default)]
-    pub is_long: bool,
-    pub quantity: FPDecimal,
-    pub entry_price: FPDecimal,
-    #[serde(default)]
-    pub effective_margin: FPDecimal,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct DerivativeLimitOrder {
-    pub order_info: OrderInfo,
-    pub order_type: i32,
-    pub margin: FPDecimal,
-    pub fillable: FPDecimal,
-    pub trigger_price: Option<FPDecimal>,
-    pub order_hash: String,
-}
-
-impl DerivativeLimitOrder {
-    pub fn new(
-        margin: FPDecimal,
-        fillable: FPDecimal,
-        order_hash: String,
-        trigger_price: Option<FPDecimal>,
-        order_type: i32,
-        order_info: OrderInfo,
-    ) -> DerivativeLimitOrder {
-        DerivativeLimitOrder {
-            margin,
-            fillable,
-            order_hash,
-            trigger_price,
-            order_type,
-            order_info,
-        }
-    }
-    pub fn is_reduce_only(&self) -> bool {
-        self.margin.is_zero()
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct DerivativePosition {
-    pub subaccount_id: String,
-    pub market_id: String,
-    pub position: Position,
-}
-
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct SubaccountEffectivePositionInMarketResponse {
     pub state: Option<EffectivePosition>,
-}
-
-#[allow(non_snake_case)]
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct TrimmedDerivativeLimitOrder {
-    pub price: FPDecimal,
-    pub quantity: FPDecimal,
-    pub margin: FPDecimal,
-    pub fillable: FPDecimal,
-    pub isBuy: bool,
-    pub order_hash: String,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -167,77 +104,4 @@ pub struct SpotMarketVolatilityResponse {
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct DerivativeMarketResponse {
     pub market: FullDerivativeMarket,
-}
-
-/// Deposit is data format for the subaccount deposit
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct Deposit {
-    #[serde(default)]
-    pub available_balance: FPDecimal,
-    #[serde(default)]
-    pub total_balance: FPDecimal,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct PerpetualMarketInfo {
-    pub market_id: String,
-    #[serde(default)]
-    pub hourly_funding_rate_cap: FPDecimal,
-    #[serde(default)]
-    pub hourly_interest_rate: FPDecimal,
-    #[serde(default)]
-    pub next_funding_timestamp: i64,
-    pub funding_interval: i64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct PerpetualMarketFunding {
-    #[serde(default)]
-    pub cumulative_funding: FPDecimal,
-    #[serde(default)]
-    pub cumulative_price: FPDecimal,
-    #[serde(default)]
-    pub last_timestamp: i64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct PerpetualMarketState {
-    pub market_info: PerpetualMarketInfo,
-    pub funding_info: PerpetualMarketFunding,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct FullDerivativeMarketPerpetualInfo {
-    pub perpetual_info: PerpetualMarketState,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct FullDerivativeMarket {
-    pub market: Option<DerivativeMarket>,
-    pub info: Option<FullDerivativeMarketPerpetualInfo>,
-    pub mark_price: FPDecimal,
-}
-
-#[allow(non_snake_case)]
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct DerivativeMarket {
-    pub ticker: String,
-    pub oracle_base: String,
-    pub oracle_quote: String,
-    #[serde(default)]
-    pub oracle_type: i32,
-    #[serde(default)]
-    pub oracle_scale_factor: u32,
-    pub quote_denom: String,
-    pub market_id: String,
-    pub initial_margin_ratio: FPDecimal,
-    pub maintenance_margin_ratio: FPDecimal,
-    pub maker_fee_rate: FPDecimal,
-    pub taker_fee_rate: FPDecimal,
-    #[serde(default)]
-    pub isPerpetual: bool,
-    #[serde(default)]
-    pub status: i32,
-    pub min_price_tick_size: FPDecimal,
-    pub min_quantity_tick_size: FPDecimal,
 }
